move multi-report back to HID.h
This makes it easier to get the simulator tests working again.
