+++ plugins/KeyboardioHID/src/HID.h
@@ -22,6 +22,7 @@
 #include <stdint.h>
 #include <Arduino.h>
 #include "HIDDefs.h"
+#include "HIDD.h"
 #include "HID-Settings.h"
 
 #if defined(USBCON)
@@ -47,6 +48,40 @@ typedef struct {
   EndpointDescriptor in;
 } HIDDescriptor;
 #pragma pack(pop)
+
+class HIDSubDescriptor {
+ public:
+  HIDSubDescriptor *next = NULL;
+  HIDSubDescriptor(const void *d, const uint16_t l)
+    : data(d), length(l) {}
+
+  const void *data;
+  const uint16_t length;
+};
+
+class HID_ : public HIDD {
+ public:
+  HID_();
+  int begin();
+  int SendReport(uint8_t id, const void *data, int len) override;
+  void AppendDescriptor(HIDSubDescriptor *node);
+  uint8_t getLEDs() {
+    return outReport[1];
+  }
+
+ protected:
+  // Implementation of the PluggableUSBModule
+  int getDescriptor(USBSetup &setup) override;
+  uint8_t getShortName(char *name);
+
+ private:
+  HIDSubDescriptor *rootNode;
+};
+
+// Replacement for global singleton.
+// This function prevents static-initialization-order-fiasco
+// https://isocpp.org/wiki/faq/ctors#static-init-order-on-first-use
+HID_ &HID();
 
 #define D_HIDREPORT(length) \
   { 9, 0x21, 0x11, 0x01, 0, 1, 0x22, lowByte(length), highByte(length) }
