+++ plugins/KeyboardioHID/src/MultiReport/AbsoluteMouse.h
@@ -29,7 +29,6 @@ THE SOFTWARE.
 #include <Arduino.h>
 #include "HID.h"
 #include "HID-Settings.h"
-#include "MultiHIDD.h"
 #include "../DeviceAPIs/AbsoluteMouseAPI.h"
 
 class AbsoluteMouse_ : public AbsoluteMouseAPI {
