+++ plugins/KeyboardioHID/src/MultiReport/SystemControl.h
@@ -30,7 +30,6 @@ THE SOFTWARE.
 #include "HID.h"
 #include "HID-Settings.h"
 #include "HIDTables.h"
-#include "MultiHIDD.h"
 #include "tusb_hid.h"
 
 #define DESCRIPTOR_SYSTEM_CONTROL(...)                  \
